+++ packages/graphql-mini-transforms/sewing-kit.config.ts
@@ -1,0 +1,8 @@
+import {createPackage, Runtime} from '@sewing-kit/config';
+
+import {quiltPackage} from '../../config/sewing-kit';
+
+export default createPackage(pkg => {
+  pkg.entry({root: './src/index'});
+  pkg.use(quiltPackage());
+});