+++ packages/graphql-typescript-definitions/src/print/index.ts
@@ -1,2 +1,4 @@
-export {printDocument, Options as PrintDocumentOptions} from './document';
-export {generateSchemaTypes, Options as PrintSchemaOptions} from './schema';
+export {printDocument} from './document';
+export type {Options as PrintDocumentOptions} from './document';
+export {generateSchemaTypes} from './schema';
+export type {Options as PrintSchemaOptions} from './schema';